Refactor, add filtered menu base class for all post factory menus.
