+++ opengever/dossier/menu.py
@@ -1,23 +1,13 @@
 from five import grok
-from opengever.dossier.behaviors.dossier import IDossierMarker
-from ftw.contentmenu.interfaces import IContentmenuPostFactoryMenu
-from zope.interface import Interface
+from opengever.base.menu import FilteredPostFactoryMenu
 from opengever.dossier import _
-from opengever.dossier.factory_utils import order_factories
+from opengever.dossier.behaviors.dossier import IDossierMarker
 from opengever.meeting import is_meeting_feature_enabled
+from zope.interface import Interface
 
 
-class DossierPostFactoryMenu(grok.MultiAdapter):
-    """If a task is added to another task, it is called subtask. So we need
-    to change the name of the task in the add-menu if we are in a task.
-    """
-
+class DossierPostFactoryMenu(FilteredPostFactoryMenu):
     grok.adapts(IDossierMarker, Interface)
-    grok.implements(IContentmenuPostFactoryMenu)
-
-    def __init__(self, context, request):
-        self.context = context
-        self.request = request
 
     def is_filtered(self, factory):
         factory_id = factory.get('id')
@@ -29,16 +19,9 @@ class DossierPostFactoryMenu(grok.MultiAdapter):
         return False
 
     def __call__(self, factories):
-        if not IDossierMarker.providedBy(self.context):
-            # use default
-            return factories
-
-        filtered_factories = []
+        factories = super(DossierPostFactoryMenu, self).__call__(factories)
         for factory in factories:
             if factory.get('id') == u'opengever.dossier.businesscasedossier':
                 factory['title'] = _(u'Subdossier')
 
-            if not self.is_filtered(factory):
-                filtered_factories.append(factory)
-
-        return order_factories(self.context, filtered_factories)
+        return factories
